requirements.txt
langchain 
langchain-core
langchain-community
langserve
uvicorn 
fastapi
requests
dotenv
google-generativeai
groq 
pydantic 
python-dotenv
chromadb 




